+++ src/main/frontend/components/PathResult.tsx
