+++ src/main/frontend/styles/tailwind.css
@@ -1,0 +1,1 @@
+/*สร้างไปงั้นแหละไม่มีไร*/
